pyproject.toml
[project]
name = "mcp-nano-banana"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "mcp[cli]>=1.13",
    "google-generativeai>=0.5.4",
    "python-dotenv>=1.0.1",
]

[project.scripts]
mcp-nano-banana = "mcp-nano-banana.main:main"

[tool.setuptools.packages.find]
where = ["src"]